+++ docs/about/release-notes.rst
@@ -43,12 +43,16 @@ Features
 ~~~~~~~~
 
 * Added new string-formatting options `#3017 <https://github.com/scipp/scipp/pull/3017>`_, `#3028 <https://github.com/scipp/scipp/pull/3028>`_.
+* Add ``size`` property to :class:`scipp.Variable` and :class:`scipp.DataArray`.
 
 Breaking changes
 ~~~~~~~~~~~~~~~~
 
 Bugfixes
 ~~~~~~~~
+
+* Fix access to value of 0-D variable with "spatial transform" dtypes using the ``value`` property.
+  Previously this raised an exception if the dtype was ``affine_transform3``, ``translation3``, or ``rotation3`` `#3033 <https://github.com/scipp/scipp/pull/3033>`_.
 
 Documentation
 ~~~~~~~~~~~~~
@@ -78,6 +82,7 @@ Features
 * Added support for arbitrary unit, degrees Celsius and other special units `#2931 <https://github.com/scipp/scipp/pull/2931>`_.
 * :class:`scipp.Dataset` now supports ``drop_coords``, which returns :class:`scipp.Dataset` without the given coordinate by names  `#2940 <https://github.com/scipp/scipp/pull/2940>`_.
 * :class:`scipp.DataArray` now supports ``drop_coords``/ ``drop_masks``/ ``drop_attrs``, which returns :class:`scipp.DataArray` without the given coordinates/masks/attributes names by names `#2940 <https://github.com/scipp/scipp/pull/2940>`_.
+* Added ``clear``, ``copy`` and ``popitem`` to the ``.coords``, ``.masks``, and ``.attrs`` accessors `#3014 <https://github.com/scipp/scipp/pull/3014>`_.
 
 Breaking changes
 ~~~~~~~~~~~~~~~~